app/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2
django-extensions==3.1.2
django-modeltranslation==0.17.1
django-picklefield==3.0.1
djangorestframework==3.12.4
easy-thumbnails==2.7.1
gunicorn==20.1.0
mistune==0.8.4
Pillow==8.2.0
psycopg2-binary==2.8.6
pygal==2.4.0
pytz==2021.1
reportlab==3.5.67
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
xlrd==2.0.1
xlwt==1.3.0
